refactor: share llm usage aggregation helpers
